Refactor Translator into separate module
