+++ main_rag.py
@@ -1,7 +1,7 @@
 # Local imports
 from python.utils import FolderTree
 from python.process import PDFProcessor
-from python.process import Translator
+from python.translation import Translator
 from python.vectorise import Chunker, Vectoriser
 from python.run import RagHTASubmission
 from python.open_ai import validate_api_key
