+++ python/run.py
@@ -10,7 +10,8 @@ if project_dir not in sys.path:
 
 # Import necessary components
 from python.utils import FolderTree, HeadingPrinter
-from python.process import PDFProcessor, Translator
+from python.process import PDFProcessor
+from python.translation import Translator
 from python.vectorise import Chunker, Vectoriser
 from python.retrieve import ChunkRetriever, PICOExtractor
 from python.open_ai import validate_api_key
